# docs/ai_api_openapi.yaml
openapi: 3.0.3
info:
  title: AI Integration Fulfillment API
  version: 1.0.0
  description: |
    Simple REST API for AI agents to create orders, manage inventory, and track shipments.
servers:
  - url: https://YOUR_HOST/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKEY
  schemas:
    InventoryItem:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        available:
          type: integer
        reserved:
          type: integer
        total_on_hand:
          type: integer
    InventoryTotalItem:
      type: object
      properties:
        item_id:
          type: string
        sku:
          type: string
        name:
          type: string
        total_on_hand:
          type: integer
        available:
          type: integer
        reserved:
          type: integer
    Product:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        weight: { type: number }
        weight_unit: { type: string }
        dimensions:
          type: object
          properties:
            length: { type: number }
            width: { type: number }
            height: { type: number }
            unit: { type: string }
    Order:
      type: object
      properties:
        order_number: { type: string }
        status: { type: string }
        tracking:
          type: object
          properties:
            carrier: { type: string }
            tracking_number: { type: string }
            url: { type: string }
    ReturnItem:
      type: object
      required: [sku, quantity]
      properties:
        sku: { type: string }
        quantity: { type: integer, minimum: 1 }
    ReturnCreateRequest:
      type: object
      required: [order_number, items]
      properties:
        created_at: { type: string, format: date-time, nullable: true }
        id: { type: string, description: "External return ID", nullable: true }
        method_code: { type: string, nullable: true }
        reason_code: { type: string, nullable: true }
        rma: { type: string, nullable: true }
        order_number: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReturnItem'
        tracking:
          type: object
          nullable: true
          properties:
            company: { type: string, nullable: true }
            number: { type: string, nullable: true }
            url: { type: string, nullable: true }
    ReturnCreateResponse:
      type: object
      properties:
        success: { type: boolean }
        return:
          type: object
          properties:
            id: { type: string }
            status: { type: string }
            order_number: { type: string }
            rma: { type: string, nullable: true }
            tracking_number: { type: string, nullable: true }
    PodProductFulfillmentCostResponse:
      type: object
      properties:
        sku: { type: string }
        fulfillment_cost: { type: number, nullable: true }
        fulfillment_cost_raw: { type: string, nullable: true }
    PodProductCostResponse:
      type: object
      properties:
        sku: { type: string }
        cost: { type: number, nullable: true }
    PodProductInventoryResponse:
      type: object
      properties:
        sku: { type: string }
        available: { type: integer }
paths:
  /inventory:
    get:
      summary: List inventory
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer }
        - in: query
          name: sku_like
          schema: { type: string }
      responses:
        '200':
          description: Inventory list
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventory:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryItem'
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
  /inventory/totals:
    get:
      summary: Get aggregated totals for client items
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: sku_like
          schema: { type: string }
      responses:
        '200':
          description: Aggregated inventory totals
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventory:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryTotalItem'
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
  /inventory/totals/{sku}:
    get:
      summary: Get aggregated totals for a single SKU
      parameters:
        - in: path
          name: sku
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Aggregated totals for SKU
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryTotalItem'
  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total: { type: integer }
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
  /pod/products/{sku}/fulfillment-cost:
    get:
      summary: Get POD product fulfillment cost by SKU
      description: Case-insensitive SKU lookup against the global products table.
      parameters:
        - in: path
          name: sku
          required: true
          schema: { type: string }
      responses:
        '200':
          description: POD product fulfillment cost
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodProductFulfillmentCostResponse'
        '404':
          description: Product not found
  /pod/products/{sku}/cost:
    get:
      summary: Get POD product cost by SKU
      description: Case-insensitive SKU lookup against the global products table.
      parameters:
        - in: path
          name: sku
          required: true
          schema: { type: string }
      responses:
        '200':
          description: POD product cost
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodProductCostResponse'
        '404':
          description: Product not found
  /pod/products/{sku}/inventory:
    get:
      summary: Get POD product available inventory by SKU
      description: Case-insensitive SKU lookup against the global products table.
      parameters:
        - in: path
          name: sku
          required: true
          schema: { type: string }
      responses:
        '200':
          description: POD product inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodProductInventoryResponse'
        '404':
          description: Product not found
  /orders:
    post:
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_number: { type: string }
                customer: { type: object }
                shipping_address: { type: object }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      sku: { type: string }
                      quantity: { type: integer }
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order_id: { type: string }
                  status: { type: string }
  /orders/{order_number}:
    get:
  /orders/{order_number}/address:
    patch:
      summary: Update shipping address for an order
      parameters:
        - in: path
          name: order_number
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                company: { type: string }
                street1: { type: string }
                street2: { type: string }
                city: { type: string }
                state: { type: string }
                postalCode: { type: string }
                country: { type: string }
                phone: { type: string }
                residential: { type: boolean }
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
  /orders/{order_number}/cancel:
    post:
      summary: Cancel an order
      parameters:
        - in: path
          name: order_number
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
  /returns:
    post:
      summary: Create a return
      description: |
        Creates a return for the authenticated client. The `order_number` is required and must belong to your store.
        Items must include `sku` and positive `quantity`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnCreateRequest'
      responses:
        '201':
          description: Return created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: boolean }
                  message: { type: string }
                  code: { type: string, example: VALIDATION_ERROR }
        '401':
          description: Unauthorized
        '500':
          description: Server error